+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Refactor `asteroid`
 - Refactor `blockade`
 - Refactor `candy-crush`
+- Refactor `coin-runner`
 
 ---